Refactoring for mobile view and visitors count
