+++ src/App.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import logo from "./logo.svg";
 import "./App.css";
 import { BrowserRouter } from "react-router-dom";
 import { Routing } from "./router/route";
 import { AuthProvider } from "./context/AuthContext/AuthProvider";
-import { Fork } from "./components/helpers/Fork";
+import { Fork } from "./components/Fork";
 
 function App() {
   return (
